+++ summoner-cli/src/Summoner/CLI.hs
@@ -166,7 +166,7 @@ data Command
 
 -- | Options parsed with @new@ command
 data NewOpts = NewOpts
-    { newOptsProjectName :: Text           -- ^ project name
+    { newOptsProjectName :: Maybe Text     -- ^ project name
     , newOptsIgnoreFile  :: Bool           -- ^ ignore all config files if 'True'
     , newOptsNoUpload    :: Bool           -- ^ don't upload to github
     , newOptsOffline     :: Bool           -- ^ Offline mode
@@ -233,7 +233,7 @@ licenseText = LicenseList <$> optional
 -- | Parses options of the @new@ command.
 newP :: Parser Command
 newP = do
-    newOptsProjectName <- strArgument (metavar "PROJECT_NAME")
+    newOptsProjectName <- optional $ strArgument (metavar "PROJECT_NAME")
     newOptsIgnoreFile  <- ignoreFileP
     newOptsNoUpload    <- noUploadP
     newOptsOffline     <- offlineP