+++ summoner-cli/CHANGELOG.md
@@ -48,7 +48,9 @@ The change log is available [on GitHub][2].
 * Remove `tasty` dependencies from tests.
 * Use `defaultDescription` for project description if not specified.
 * [#239](https://github.com/kowainik/summoner/issues/239):
-   Make categories not mandatory.
+  Make categories not mandatory.
+* [#243](https://github.com/kowainik/summoner/issues/243):
+  Make project name optional in CLI arguments.
 
 
 ## 1.1.0.1 – Sep 10, 2018